@@ -1,8 +1,8 @@
 # Autogenerated by devscripts/update-version.py
 
-__version__ = '2023.11.14'
+__version__ = '2023.11.16'
 
-RELEASE_GIT_HEAD = 'a9d3f4b20a3533d2a40104c85bc2cc6c2564c800'
+RELEASE_GIT_HEAD = '24f827875c6ba513f12ed09a3aef2bbed223760d'
 
 VARIANT = None
 
@@ -12,4 +12,4 @@ CHANNEL = 'stable'
 
 ORIGIN = 'yt-dlp/yt-dlp'
 
-_pkg_version = '2023.11.14'
+_pkg_version = '2023.11.16'
